backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pandas==2.1.3
numpy==1.26.2
matplotlib==3.8.2
Pillow==10.1.0
openpyxl==3.1.2
requests==2.31.0
python-multipart==0.0.6
pytz==2023.3